main

+++ routes/campgrounds.js
@@ -1,20 +1,8 @@
 const express = require("express");
 const router = express.Router();
 const catchAsync = require("../utils/catchAsync");
-const ExpressError = require("../utils/ExpressError");
 const Campground = require("../models/campground");
-const { campgroundSchema } = require("../schemas.js");
-const { isLoggedIn } = require("../middleware");
-
-const validateCampground = (req, res, next) => {
-  const { error } = campgroundSchema.validate(req.body);
-  if (error) {
-    const msg = error.details.map((el) => el.message).join(",");
-    throw new ExpressError(msg, 400);
-  } else {
-    next();
-  }
-};
+const { isLoggedIn, isAuthor, validateCampground } = require("../middleware");
 
 router.get(
   "/",
@@ -34,6 +22,7 @@ router.post(
   catchAsync(async (req, res, next) => {
     // if (!req.body.campground) throw new Express("Invalid Campground data", 400);
     const campground = new Campground(req.body.campground);
+    campground.author = req.user._id;
     await campground.save();
     req.flash("success", "新增完成!");
     res.redirect(`/campgrounds/${campground._id}`);
@@ -43,7 +32,9 @@ router.get(
   "/:id",
   catchAsync(async (req, res) => {
     const { id } = req.params;
-    const c = await Campground.findById(id).populate("reviews");
+    const c = await Campground.findById(id)
+      .populate({ path: "reviews", populate: { path: "author" } })
+      .populate("author");
     if (!c) {
       req.flash("error", "找不到指定的露營地");
       return res.redirect("/campgrounds");
@@ -54,10 +45,11 @@ router.get(
 router.get(
   "/:id/edit",
   isLoggedIn,
+  isAuthor,
   catchAsync(async (req, res) => {
     const campground = await Campground.findById(req.params.id);
     if (!campground) {
-      req.flash("error", "編輯失敗");
+      req.flash("error", "無法找到此露營地");
       return res.redirect("/campgrounds");
     }
     res.render("campgrounds/edit", { campground });
@@ -66,10 +58,12 @@ router.get(
 router.put(
   "/:id",
   isLoggedIn,
+  isAuthor,
   validateCampground,
   catchAsync(async (req, res) => {
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(
+
+    const camp = await Campground.findByIdAndUpdate(
       id,
       {
         ...req.body.campground,
@@ -77,15 +71,21 @@ router.put(
       { new: true }
     );
     req.flash("success", "編輯成功!");
-    res.redirect(`/campgrounds/${campground._id}`);
+    res.redirect(`/campgrounds/${camp._id}`);
   })
 );
 
 router.delete(
   "/:id",
   isLoggedIn,
+  isAuthor,
   catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground.author.equals(req.user._id)) {
+      req.flash("error", "你沒有刪除權限!!");
+      return res.redirect(`/campgrounds/${id}`);
+    }
     await Campground.findByIdAndDelete(id);
     req.flash("success", "已刪除露營地");
     res.redirect("/campgrounds");
